Fixed bug in MaterialUI where buttons wouldn't show as disabled if they were set to non-interactable when the scene was first initialised.
Since lastInteractableState is initialised to default(bool), which is
false, if a button started as non-interactable, it would never get
updated correctly by ButtonInteractableControl. Usually this wouldn't be
an issue since ButtonInteractableControl also runs in the editor, but if
the button was enabled in the editor and then disabled by a script in
the same frame as the scene was loaded this could cause an issue.
Also moved member initialisation out of OnEnable, since this is only
called if the object is enabled but Update() is still called on disabled
scripts that have the ExecuteInEditMode attribute (this can cause
NullReferenceExceptions if the script is disabled in the scene and you
click on it in the inspector, since it'll try to call Update() but
OnEnable() may never have been called).