Fixed issue that was causing screens to sometimes be positioned in the wrong place.
If a user triggered a transition to a certain screen twice in quick
succession, without waiting for the first transition to complete, the
second screen would be re-positioned around the last click location.
This was fixed simply by preventing the current screen from being set
while a transition is in progress.